libshvbroker: Refactor LdapConfig
